Serialization (#15)
* Add Routes property to Daytype class with JsonIgnore

Added a using directive for System.Text.Json.Serialization.
Introduced a new property Routes to the Daytype class, marked
with the [JsonIgnore] attribute to exclude it from JSON
serialization.

* Remove unnecessary using directive from Direction.cs

The `using System.Diagnostics.CodeAnalysis;` directive has been removed from the `Direction.cs` file. This directive was likely used for code analysis and suppression of warnings, but it has been deemed unnecessary and thus removed. The rest of the file, including the namespace and class definition, remains unchanged.